Banking Application
Hello Guys, This is a Banking Application where the user/Employee of a Bank can choose what to do like depositing the amount or creating the user's account. Also to do it he has to log in first. He also has access to remove the data and store it. The display of the data stored will be shown just like the database(DBMS).